feat(app_modal_page): add a file dedicatedto constant MenuItem
